feat: refactor auths
