+++ apps/web/@/messages/en.json
@@ -165,7 +165,8 @@
       "sign_in_with_linkedin": "Sign in with LinkedIn",
       "sign_in_with_instagram": "Sign in with Instagram",
       "sign_in_with_youtube": "Sign in with YouTube",
-      "sign_in_with_tiktok": "Sign in with TikTok"
+      "sign_in_with_tiktok": "Sign in with TikTok",
+      "sign_in_to_your_account_to_continue": "Sign in to your account to continue"
     },
     "sign_up": {
       "title": "Sign Up",
