+++ apps/web/@/molecules/auth/forgot-password/index.tsx
@@ -1,0 +1,3 @@
+export default function ForgotPassword() {
+  return <div>ForgotPassword</div>
+}
