+++ apps/web/@/molecules/auth/reset-password/index.tsx
@@ -1,0 +1,3 @@
+export default function ResetPassword() {
+  return <div>ResetPassword</div>
+}
